containers/load_monthly_data/requirements.txt
google-cloud-storage
google-cloud-bigquery==1.23.0

pytest
mock